[bugfix] Keep the same note in view after orientation change
Removing the listview's cacheBuffer setting took care of that.
The cacheBuffer was wrong for other reasons (it chouldn't cache Note
objects because those contain an editor that might have old contents)
and removing it has the side effect of not tempting the ListView to
choose the wrong Note to display after an orientation change.